Move parse orchestration to stages/parsing.py; parsers/ is now pure plugins
stages/ now contains the full ordered pipeline (parsing -> toc -> chunking ->
summarize), making the processing flow legible in one place. The router logic
(content-type detection, parser selection, plaintext handling) moves from
parsers/router.py into stages/parsing.py.

parsers/ is now a pure implementation family (docling, liteparse) with no
orchestration — stages/parsing.py decides which concrete parser to call.

- Delete rag/parsers/router.py
- Add rag/stages/parsing.py
- Update imports in indexing.py and __init__.py
